Put the erasure proofs where they run on every gate, and keep the one that binds a port
Main split the Gateway suite: CcDirector.Gateway.Tests is now PARKED behind -Parked because
it serializes machine-wide, and the pure tests moved to CcDirector.Gateway.UnitTests, which
runs in the default two-minute gate. Every acceptance proof this work item added was sitting
in the parked project, so an ordinary gate would have gone green without executing one of them.

Judged one file at a time against the split's own rule - construct no host, bind no port,
spawn no process, touch no live database, mutate no process-global environment variable:

  MOVED to the fast suite, so they run on every gate
    History/SessionHistoryStoreTests.cs                        main had already moved this file
    History/AnInFlightWriterIsRefusedByTheWatermarkTests.cs    store over a temp SQLite harness
    History/TheErasureIsJudgedByMaterialAgeTests.cs            two store instances, one file
    History/TheErasureCarriesTheTenantPredicateOnNpgsqlTests.cs  never opens a connection
    History/TheDeletionBoundaryGuardsTests.cs                  five of its six facts

  STAYED behind the lock
    Prompts/PromptDeleteErasesTheDerivedCopyTests.cs   real HTTP pipeline, binds a socket
    Prompts/PromptEndpointsTests.cs                    host-bound; main kept it here
    OmittedTenantBoundaryFailClosedTests.cs            mutates process-global environment variables
    History/TheSealEndpointRefusesAPreErasureSessionTests.cs   split out of the guards file

The last one is the deliberate part. TheDeletionBoundaryGuardsTests held six facts and exactly
one of them starts a WebApplication and binds a socket. Leaving the file whole would have parked
five proofs that need nothing machine-global, so the host-bound fact moved into its own file here
and the other five went to the fast suite. The guards file names where the sixth went, because a
list of guards that quietly drops a guard reads as complete.

One sentence was corrected in the fact that moved: it said admission compares the watermark
against the session's own start. It does not - it compares against the first moment this Gateway
saw the session - and relocating a false sentence by hand would be writing it fresh.

Three files the rebase resurrected are deleted: main renamed the History directory while this
branch had renamed three files inside it, so the pre-rename copies came back alongside their
replacements. They compile, which is why this is worth naming: the duplicates carried the exact
test names an inspection rejected.

Build: whole solution, Debug, zero warnings, zero errors.
