+++ src/CcDirector.Gateway.UnitTests/History/TheDeletionBoundaryGuardsTests.cs
@@ -1,5 +1,3 @@
-using System.Net;
-using System.Net.Http.Json;
 using CcDirector.AgentBrain;
 using CcDirector.Core.Tenancy;
 using CcDirector.Gateway.Contracts;
@@ -7,9 +5,7 @@ using CcDirector.Gateway.Data.Entities;
 using CcDirector.Gateway.History;
 using CcDirector.Gateway.Prompts;
 using CcDirector.Gateway.Tests.Data;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
 using Xunit;
 
 namespace CcDirector.Gateway.Tests.History;
@@ -27,6 +23,11 @@ namespace CcDirector.Gateway.Tests.History;
 /// (in PromptEndpoints) says so: the cross-process file append and the roll-up insert are deliberately
 /// contained rather than defeated. What is proved here is what each guard DOES, one fact at a time - and
 /// a test name is a claim like any other sentence in this product.
+///
+/// A SIXTH FACT USED TO LIVE HERE and is now in <c>CcDirector.Gateway.Tests</c>, as
+/// <c>TheSealEndpointRefusesAPreErasureSessionTests</c>: it drives the seal refusal over real HTTP, so it
+/// binds a socket and belongs behind the machine-wide lock. It is named here rather than left to be
+/// noticed, because a guard list that quietly drops one of its guards reads as complete.
 /// </summary>
 public sealed class TheDeletionBoundaryGuardsTests : IDisposable
 {
@@ -211,69 +212,6 @@ public sealed class TheDeletionBoundaryGuardsTests : IDisposable
         store.SaveRollup("thefrederiksen/devthrottle", day, "A later paragraph.", "hash2", 0,
             DateTime.UtcNow, DateTime.UtcNow.AddSeconds(1));
         Assert.Equal("A later paragraph.", Assert.Single(store.ReadRollups(day, day)).SummaryText);
-    }
-
-    /// <summary>
-    /// FINDING 3, driven through the REAL endpoint rather than the store. The seal request carries no
-    /// material time, and the endpoint used to substitute the moment the request ARRIVED - which is always
-    /// newer than an erasure that already happened, so every seal was admitted after every delete. The
-    /// previous test passed a backdated value the endpoint never produces, which is why it passed over a
-    /// live hole.
-    ///
-    /// There is now no time to pass: the store compares the watermark against the SESSION'S OWN START. This
-    /// fact goes over HTTP so the thing under test is the path a session actually uses.
-    /// </summary>
-    [Fact]
-    public async Task A_seal_arriving_after_the_delete_is_refused_by_the_real_endpoint()
-    {
-        var db = _harness.Open();
-        var store = new SessionHistoryStore(db);
-        var startedBeforeTheDelete = DateTime.UtcNow.AddHours(-2);
-        store.UpsertLive("dir-1", Session(created: startedBeforeTheDelete), startedBeforeTheDelete);
-
-        var builder = WebApplication.CreateBuilder();
-        builder.Logging.ClearProviders();
-        var app = builder.Build();
-        app.Urls.Add("http://127.0.0.1:0");
-        HistoryEndpoints.Map(app, store, tenantBoundary: null);
-        await app.StartAsync();
-        using var client = new HttpClient { BaseAddress = new Uri(app.Urls.First()) };
-
-        try
-        {
-            store.ErasePromptDerived();
-
-            var response = await client.PostAsJsonAsync("/history/sessions/s1/summary", new SealSessionSummaryRequest
-            {
-                Summary = "A farewell composed from the conversation the member just erased.",
-                WhatWasBuilt = new[] { "something" },
-            });
-
-            Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            using var ctx = db.CreateContext();
-            var row = ctx.SessionHistory.AsNoTracking().Single(e => e.SessionId == "s1");
-            Assert.Null(row.SummaryText);
-            Assert.Null(row.SummaryKind);
-
-            // A CALLER-CONTROLLED START NO LONGER BUYS ADMISSION. This session claims it began after the
-            // erasure - the exact value the previous rule trusted - but this Gateway first saw it now, and
-            // "now" is after the erasure, so it seals. The point of the assertion below is not that it
-            // succeeds but that the value deciding it is OURS: see the refusal fact that follows.
-            store.UpsertLive("dir-1", Session(id: "s2", created: DateTime.UtcNow.AddSeconds(1)), DateTime.UtcNow);
-            var ok = await client.PostAsJsonAsync("/history/sessions/s2/summary", new SealSessionSummaryRequest
-            {
-                Summary = "A farewell for work that started after the delete.",
-            });
-            Assert.Equal(HttpStatusCode.OK, ok.StatusCode);
-            using var after = db.CreateContext();
-            Assert.Equal("A farewell for work that started after the delete.",
-                after.SessionHistory.AsNoTracking().Single(e => e.SessionId == "s2").SummaryText);
-        }
-        finally
-        {
-            await app.StopAsync();
-            await app.DisposeAsync();
-        }
     }
 
     /// <summary>
